+++ packages/source-trpc/src/index.ts
@@ -1,0 +1,7 @@
+export { fromTrpc } from './from-trpc'
+export type {
+  FromTrpcOptions,
+  TrpcCreateContextInput,
+  TrpcProcedureKind,
+  TrpcToolResult,
+} from './types'